# api/agent/swagger.yaml
basePath: /v1
definitions:
...
  api.runStep:
    properties:
      error:
        type: string
      finished_at:
        type: string
      id:
        type: string
      input:
        additionalProperties: {}
        type: object
      output:
        additionalProperties: {}
        type: object
      started_at:
        type: string
      status:
        type: string
    type: object
  api.runSummary:
    properties:
      created_at:
        type: string
      error:
        type: string
      finished_at:
        type: string
      id:
        type: string
      inputs:
        additionalProperties: {}
        type: object
      outputs:
        additionalProperties: {}
        type: object
      started_at:
        type: string
      status:
        type: string
      steps:
        items:
          $ref: '#/definitions/api.runStep'
        type: array
      workflow_id:
        type: string
      workflow_version:
        type: integer
    type: object
  api.runWorkflowRequest:
    properties:
      bindings:
        additionalProperties:
          type: string
        type: object
      inputs:
        additionalProperties: {}
        type: object
    type: object
  api.workflowBindingDetail:
    properties:
      connector_type:
        description: |-
          ConnectorType is empty when it could not be inferred for every step
          that references this binding name, or when different steps referring
          to the same name disagree on it.
        type: string
      name:
        type: string
    type: object
  api.workflowDetail:
    properties:
      bindings:
        description: |-
          Bindings is every distinct connector binding name referenced by Steps
          (via a "${{ bindings.<name> }}" expression), each paired with its
          inferred connector type when one could be — see bindingConnectorType.
        items:
          $ref: '#/definitions/api.workflowBindingDetail'
        type: array
      id:
        type: string
      inputs:
        items:
          $ref: '#/definitions/api.workflowInputDetail'
        type: array
      next_run_at:
        description: |-
          NextRunAt is this workflow's next scheduled firing, populated from
          the schedules table (internal/scheduler) rather than parsed from the
          definition — nil for a "manual" trigger, or if the schedules row
          hasn't caught up yet with a just-installed version.
        type: string
      schema_version:
        type: integer
      source:
        type: string
      steps:
        items:
          $ref: '#/definitions/api.workflowStepDetail'
        type: array
      trigger_cron:
        description: |-
          TriggerCron and TriggerOnMissed are only set when TriggerType is
          "schedule" — see internal/workflow.Trigger (ADR-0035).
        type: string
      trigger_on_missed:
        type: string
      trigger_type:
        type: string
      version:
        type: integer
    type: object
  api.workflowInputDetail:
    properties:
      default: {}
      description:
        type: string
      enum:
        items:
          type: string
        type: array
      name:
        type: string
      required:
        type: boolean
      type:
        type: string
    type: object
  api.workflowStepDetail:
    properties:
      binding_name:
        description: |-
          BindingName is the logical name this step's Connector expression
          refers to (the "db" in "${{ bindings.db }}"), populated only when
          Connector is exactly that shape — see bindingName.
        type: string
      connector:
        type: string
      connector_type:
        description: |-
          ConnectorType is the connector type inferred for BindingName, so a
          client can offer a <select> of matching connectors instead of a
          free-text id — see bindingConnectorType. Left empty when it can't be
          inferred (no owning plugin found, or that plugin declares more than
          one connector type).
        type: string
      id:
        type: string
      uses:
        type: string
      with:
        additionalProperties: {}
        type: object
    type: object
  api.workflowSummary:
    properties:
      id:
        type: string
      installed_at:
        type: string
      version:
        type: integer
    type: object
info:
  contact: {}
  description: |-
    Public HTTP API for the Patchcord agent (vision document, section 10.1). Covers workflow triggering and run observability, connector CRUD and testing, application hosting, and a read-only plugin catalog listing; the rest of the vision document's API surface (actions, full plugin management) is not implemented yet.
    Every route marked with a lock icon below requires "Authorization: Bearer <admin token>" — but only once at least one admin token has been created (`patchcord auth token create`); a fresh agent answers every request unauthenticated, exactly as before this existed (ADR-0036).
  title: Patchcord Agent API
  version: "1"
paths:
...
  /workflows:
    get:
      description: Returns every installed workflow version, most recently installed
        first. The same workflow id can appear more than once if multiple versions
        are installed (workflows are immutable once published — installing a new version
        never replaces an old one).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.workflowSummary'
            type: array
      security:
      - BearerAuth: []
      summary: List installed workflows
      tags:
      - workflows
  /workflows/{id}:
    get:
      description: 'Returns one workflow version''s full definition: its steps and
        raw YAML source. Defaults to the latest installed version.'
      parameters:
      - description: Workflow id
        in: path
        name: id
        required: true
        type: string
      - description: Workflow version (defaults to the latest installed one)
        in: query
        name: version
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.workflowDetail'
        "400":
          description: version is not a valid integer
          schema:
            type: string
        "404":
          description: workflow not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get a workflow version's definition
      tags:
      - workflows
  /workflows/{id}/run:
    post:
      consumes:
      - application/json
      description: Starts a new run of the latest installed version of the named workflow
        and returns immediately with its id and initial status — it does not wait
        for any step to execute. Watch GET /runs/{id}/events or poll GET /runs/{id}
        for progress and the final result.
      parameters:
      - description: Workflow id
        in: path
        name: id
        required: true
        type: string
      - description: Inputs and connector bindings (both optional, default to none)
        in: body
        name: body
        schema:
          $ref: '#/definitions/api.runWorkflowRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.runSummary'
        "400":
          description: malformed request body, or inputs don't satisfy the workflow's
            declared schema
          schema:
            type: string
        "404":
          description: workflow not found
          schema:
            type: string
        "500":
          description: no action executor configured, or a persistence failure
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Trigger a workflow run
      tags:
      - workflows
securityDefinitions:
  BearerAuth:
    description: Admin token, if any has been created (see the top-level description).
      Pass as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"